move middleware to top
