+++ src/application/module/books/book.routes.ts
@@ -1,7 +1,7 @@
 import BookService from "./book.service";
 import BookController from "./book.controller";
 import AppRoutes from "../../common/interfaces/AppRoutes";
-import JwtMiddleware from "../../middleware/jwt.middleware";
+import JwtMiddleware from "../../../middleware/jwt.middleware";
 
 export default class BookRoutes extends AppRoutes {
   constructor() {
