+++ src/middleware/jwt.middleware.ts
@@ -1,8 +1,8 @@
 import jwt, { JwtPayload } from "jsonwebtoken";
 import { NextFunction, Request, Response } from "express";
-import { IUser } from "../entity/IUser";
-import { AuthenticatedRequest } from "../common/types/AuthenticateRequest";
-import envConfig from "../../config/dotenv.confiq";
+import envConfig from "../config/dotenv.confiq";
+import { AuthenticatedRequest } from "../application/common/types/AuthenticateRequest";
+import { IUser } from "../application/entity/IUser";
 
 export default class JwtMiddleware {
   static verifyToken(
@@ -11,11 +11,13 @@ export default class JwtMiddleware {
     next: NextFunction
   ): void | Response {
     const authHeader = request.header("Authorization");
+
     if (!authHeader || !authHeader.startsWith("Bearer ")) {
       return response
         .status(401)
         .json({ message: "Access Denied: No valid token provided" });
     }
+
     const accessToken = authHeader.replace("Bearer ", "").trim();
 
     try {
@@ -23,9 +25,11 @@ export default class JwtMiddleware {
         accessToken,
         envConfig.JwtSectret
       ) as JwtPayload;
+
       if (!payload.id || !payload.email || !payload.name) {
         return response.status(401).json({ message: "Invalid token payload" });
       }
+
       (request as AuthenticatedRequest).user = payload as IUser;
       next();
     } catch (error) {
